Restructured libvectormap and updated dependent files
