+++ ros/src/common/libs/vectormap_lib/CHANGELOG.rst
