+++ ros/src/common/libs/vectormap_lib/include/vectormap_lib/Math.h
